# api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  models.AddUserInterestRequest:
    properties:
      interest_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.AddUserPreferenceRequest:
    properties:
      Preference_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  models.DeleteUserInterestRequest:
    properties:
      interest_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.DeleteUserPreferenceRequest:
    properties:
      Preference_id:
        minimum: 4
        type: integer
      user_id:
        type: integer
    type: object
  models.EditUserInterestRequest:
    properties:
      interest_id:
        minimum: 4
        type: integer
      new_interest_name:
        minLength: 4
        type: string
      user_id:
        type: integer
    type: object
  models.EditUserPreferenceRequest:
    properties:
      Preference_id:
        type: integer
      new_Preference_name:
        minLength: 4
        type: string
      user_id:
        type: integer
    type: object
...
  models.Preferences:
    properties:
      id:
        type: integer
      preference_name:
        minLength: 4
        type: string
    type: object
...
  models.User:
    properties:
      name:
        type: string
      user_id:
        type: integer
    type: object
...
  models.UserUpdateDetails:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      user_name:
        type: string
    required:
    - email
    - password
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:3000
info:
  contact: {}
  description: SOCIAL MATCH API Gateway server.
  termsOfService: http://swagger.io/terms/
  title: Social-Match
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get All Users
      tags:
      - users
  /user/{user_id}/interests:
    get:
      description: Get a list of user interests
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User interests
          schema:
            items:
              items:
                type: string
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get user interests
      tags:
      - users
  /user/{user_id}/preferences:
    get:
      description: Get a list of user preferences
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User preferences
          schema:
            items:
              items:
                $ref: '#/definitions/models.Preferences'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get user preferences
      tags:
      - users
  /user/add-interest:
    post:
      consumes:
      - application/json
      description: Add user interest
      parameters:
      - description: Add interest request
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/models.AddUserInterestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add User Interest
      tags:
      - users
  /user/add-preference:
    post:
      consumes:
      - application/json
      description: Add user preference
      parameters:
      - description: Add preference request
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/models.AddUserPreferenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add User Preference
      tags:
      - users
...
  /user/delete-interest:
    delete:
      consumes:
      - application/json
      description: Delete user interest
      parameters:
      - description: Delete interest request
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/models.DeleteUserInterestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete User Interest
      tags:
      - users
  /user/delete-preference:
    delete:
      consumes:
      - application/json
      description: Delete user preference
      parameters:
      - description: Delete preference request
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/models.DeleteUserPreferenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete User Preference
      tags:
      - users
  /user/edit:
    put:
      consumes:
      - application/json
      description: Edit user details
      parameters:
      - description: Edit details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateDetails'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Edit User Details
      tags:
      - users
  /user/edit-interest:
    put:
      consumes:
      - application/json
      description: Edit user interest
      parameters:
      - description: Edit interest request
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/models.EditUserInterestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Edit User Interest
      tags:
      - users
  /user/edit-preference:
    put:
      consumes:
      - application/json
      description: Edit user preference
      parameters:
      - description: Edit preference request
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/models.EditUserPreferenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Edit User Preference
      tags:
      - users
...
  /user/profile:
    patch:
      consumes:
      - multipart/form-data
      description: Update user's profile photo
      parameters:
      - collectionFormat: csv
        description: Profile photos
        in: formData
        items:
          type: file
        name: image_paths
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update Profile Photo
      tags:
      - users
...
securityDefinitions:
  AdminAutherisation:
    in: header
    name: AccessToken
    type: apiKey
  UserAuthorization:
    in: header
    name: AccessToken
    type: apiKey
  UserConfirmToken:
    in: header
    name: ConfirmToken
    type: apiKey
swagger: "2.0"